monocypher-sys: use pre-generated bindings & bundle monocypher source
Generating bindings ahead of time and shipping them can help reduce
build times quite a bit (as bindgen doesn't need to be built) and
removes the libclang dependency (needed for bindgen).

As long as the API we're targeting is fairly stable, there shouldn't be
a big downside. One concern would be ensuring the generated bindings are
kept in sync. This is something we could handle with some CI
configuration.

Bundling/vendoring the source code enables very easy use in rust
programs. One case we want to be cautious about is if, for some reason,
and external monocypher lib is desired. As far as I can tell, this is
somewhat uncommon (monocypher does not appear packaged on many distros).
That said: using some features (in a similar way to openssl-sys) could
allow supporting the non-vendored/bundled case in the future.